separate out app/Server.hs and app/Main.hs
